GoSwagger
These additions allow for a swagger spec to be used with gojsonschema.  Given an
HTTP request to a URL some basic pattern patching is done and the body parameter
is compared against the body parameter in the swagger spec.  This currently
supports more complex gojsonschema constructs like allOf and anyOf even though
swaggerUI itself does not.

The final resting place of this code was meant to be elsewhere but we are putting
this development on hold as local validation is no longer a priority in our tests.