Added option to check format also on numbers
In some cases i would like to check if a entered number is valid using
my own code, for example, if i want to check if a given property points
to valid id on a database.

This patches breaks backward compatibility because so far the format
property was only used on string values so the parameter to IsFormat()
was also a string. We have to convert it to be an interface{} so we
are able to send any type of input to the function.